pipeline/archive/v1/requirements.txt
# Database
psycopg2-binary>=2.9.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# AI
anthropic>=0.18.0

# Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Document generation
python-docx>=0.8.11

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# JSON schema validation
jsonschema>=4.17.0
